move ui store types
